requirements.txt
python-telegram-bot[job-queue]>=21.0
faster-whisper>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
